feat(course-card): convert to Glimmer component and unify course access
Refactor CourseCard to a Glimmer component with typed args and computed
properties for course, currentUser, and lastUsedRepository.

Update template to consistently access course and user data via the
component instance (`this.course`), improving readability and binding
accuracy.

Implement fallback logic for last used repository, prioritizing last
pushed then last created repository for the current user.

Remove comment lines suppressing Glint errors in catalog-loading to
clean up the template.

These changes improve type safety, maintainability, and correctness
of course data rendering and user-specific repository state.
