--- app/components/course-card/difficulty-label.js
@@ -1,3 +1,0 @@
-import Component from '@glimmer/component';
-
-export default class DifficultyLabel extends Component {}
