+++ app/templates/catalog-loading.hbs
@@ -7,7 +7,6 @@
 
   <div class="grid grid-cols-1 gap-3 md:grid-cols-2 lg:grid-cols-2 mb-8">
     {{#each (repeat 6)}}
-      {{! @glint-expect-error }}
       <CourseCard />
     {{/each}}
   </div>
@@ -18,7 +17,6 @@
 
   <div class="grid grid-cols-1 gap-3 md:grid-cols-2 lg:grid-cols-3 mb-8">
     {{#each (repeat 6)}}
-      {{! @glint-expect-error }}
       <CourseCard />
     {{/each}}
   </div>
