+++ app/templates/catalog.hbs
@@ -16,7 +16,6 @@
 
   <div class="grid grid-cols-1 gap-3 md:grid-cols-2 lg:grid-cols-2 mb-8">
     {{#each this.orderedCourses as |course|}}
-      {{! @glint-expect-error: not ts-ified yet }}
       <CourseCard @course={{course}} />
     {{/each}}
   </div>
